Edu's changes to accept image-processing class and to threshold scans
Modified __init__ adding a setupdacscan function to properly set the number of frames to acquire in a DASscan. When acquirin dac scans the number of frames to acquire always reads one. Therefore, with the old approach the MERLIN class gets one image and then leaves the rest on the buffer.
Modified Frame class to accept my merlin-image-processing class. This has some extra functionality (e.g. gets the individual chip matrices) and computes sums. IT is still giving me an error when it comes to raw data...